+++ cmd/traffic/cmd/manager/service.go
@@ -377,11 +377,12 @@ func (m *Manager) UpdateIntercept(ctx context.Context, req *rpc.UpdateInterceptR
 				defer cancel()
 				var resp *systema.CreateDomainResponse
 				resp, err = sa.CreateDomain(tc, &systema.CreateDomainRequest{
-					InterceptId:    intercept.Id,
-					DisplayBanner:  action.AddPreviewDomain.DisplayBanner,
-					InterceptSpec:  intercept.Spec,
-					Host:           action.AddPreviewDomain.Ingress.L5Host,
-					PullRequestUrl: action.AddPreviewDomain.PullRequestUrl,
+					InterceptId:       intercept.Id,
+					DisplayBanner:     action.AddPreviewDomain.DisplayBanner,
+					InterceptSpec:     intercept.Spec,
+					Host:              action.AddPreviewDomain.Ingress.L5Host,
+					PullRequestUrl:    action.AddPreviewDomain.PullRequestUrl,
+					AddRequestHeaders: action.AddPreviewDomain.AddRequestHeaders,
 				})
 				if err != nil {
 					err = errors.Wrap(err, "systema: create domain")